Moved sample components to another folder, base file for main app
