+++ front-sample/src/App.js
@@ -2,6 +2,7 @@ import React, { Component } from 'react';
 import { BrowserRouter as Router, Route } from "react-router-dom";
 import './App.css';
 import FrontSample from './FrontSample';
+import Candystore from './Candystore';
 
 //let userinfo2 = UserInfov2('localhost:3333','/users/id/1');
 
@@ -10,7 +11,10 @@ class App extends Component {
     return(
       <div>
         <Router>
-          <Route path="/sample" component={FrontSample}/>
+          <div>
+            <Route path="/" exact={true} component={Candystore}/>
+            <Route path="/sample" component={FrontSample}/>
+          </div>
         </Router>
       </div>
     )
