+++ front-sample/src/Candystore.js
@@ -1,0 +1,13 @@
+import React, { Component } from 'react';
+
+class Candystore extends Component {
+  render() {
+    return(
+      <div>
+        Main app comes here
+      </div>
+    )
+  }
+}
+
+export default Candystore;
